requirements.txt
nltk==3.8.1
python-docx==0.8.11
PyPDF2==3.0.1
streamlit==1.28.0
